Rename: notes 패키지를 domain 패키지 내부로 이동시킴
