+++ src/main/java/com/code2am/stocklog/domain/notes/dao/NotesDAO.java
@@ -1,6 +1,6 @@
-package com.code2am.stocklog.notes.dao;
+package com.code2am.stocklog.domain.notes.dao;
 
-import com.code2am.stocklog.notes.models.vo.NotesVo;
+import com.code2am.stocklog.domain.notes.models.vo.NotesVo;
 import org.apache.ibatis.annotations.Mapper;
 
 import java.util.List;
