+++ src/main/java/com/code2am/stocklog/domain/notes/repository/NotesRepository.java
@@ -1,6 +1,6 @@
-package com.code2am.stocklog.notes.repository;
+package com.code2am.stocklog.domain.notes.repository;
 
-import com.code2am.stocklog.notes.models.entity.Notes;
+import com.code2am.stocklog.domain.notes.models.entity.Notes;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
